fix: adjust graphql updates
