--- schemas/objects/customProductOption/height.ts
@@ -1,9 +1,0 @@
-import {defineField} from 'sanity'
-
-export default defineField({
-  name: 'customProductOption.height',
-  title: 'Height',
-  type: 'number',
-  description: 'In mm',
-  validation: (Rule) => Rule.required().precision(2),
-})
