+++ schemas/objects/global/linkExternal.ts
