+++ schemas/objects/global/links.ts
