+++ schemas/objects/hotspot/imageWithProductHotspots.ts
