+++ schemas/objects/hotspot/productHotspots.tsx
@@ -1,6 +1,6 @@
 import {defineField} from 'sanity'
 
-import ProductTooltip from '../../components/hotspots/ProductTooltip'
+import ProductTooltip from '../../../components/hotspots/ProductTooltip'
 
 export default defineField({
   name: 'productHotspots',
