+++ schemas/objects/hotspot/spot.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import {defineField} from 'sanity'
 
-import ShopifyDocumentStatus from '../../components/media/ShopifyDocumentStatus'
+import ShopifyDocumentStatus from '../../../components/media/ShopifyDocumentStatus'
 
 export default defineField({
   name: 'spot',
+  title: 'Spot',
   type: 'object',
   fieldsets: [{name: 'position', options: {columns: 2}}],
   fields: [
