+++ schemas/objects/seo/seo.ts
