+++ schemas/objects/shopify/inventory.ts
