+++ schemas/objects/shopify/priceRange.ts
