+++ schemas/objects/shopify/productWithVariant.tsx
@@ -3,9 +3,9 @@ import pluralize from 'pluralize-esm'
 import React from 'react'
 import {defineField} from 'sanity'
 
-import ShopifyDocumentStatus from '../../components/media/ShopifyDocumentStatus'
-import {SANITY_API_VERSION} from '../../constants'
-import {getPriceRange} from '../../utils/getPriceRange'
+import ShopifyDocumentStatus from '../../../components/media/ShopifyDocumentStatus'
+import {SANITY_API_VERSION} from '../../../constants'
+import {getPriceRange} from '../../../utils/getPriceRange'
 
 export default defineField({
   name: 'productWithVariant',
