+++ schemas/objects/shopify/shopifyCollection.ts
