+++ schemas/objects/shopify/shopifyCollectionRule.tsx
