+++ schemas/objects/shopify/shopifyProductVariant.ts
